Joaoivosite/requirements.txt
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
Pillow==10.3.0
plotly==5.20.0
streamlit==1.32.2
